Changed directory structure and renamed
