+++ requirements.txt
@@ -3,3 +3,8 @@ IPy
 ipwhois
 termcolor
 ipaddress
+connexion == 2.6.0
+swagger-ui-bundle
+python_dateutil == 2.6.0
+setuptools >= 21.0.0
+redis
