--- src/api/server/.swagger-codegen/VERSION
@@ -1,1 +1,0 @@
-3.0.19
