--- src/api/server/README.md
@@ -1,49 +1,0 @@
-# Swagger generated server
-
-## Overview
-This server was generated by the [swagger-codegen](https://github.com/swagger-api/swagger-codegen) project. By using the
-[OpenAPI-Spec](https://github.com/swagger-api/swagger-core/wiki) from a remote server, you can easily generate a server stub.  This
-is an example of building a swagger-enabled Flask server.
-
-This example uses the [Connexion](https://github.com/zalando/connexion) library on top of Flask.
-
-## Requirements
-Python 3.5.2+
-
-## Usage
-To run the server, please execute the following from the root directory:
-
-```
-pip3 install -r requirements.txt
-python3 -m swagger_server
-```
-
-and open your browser to here:
-
-```
-http://localhost:8888/v1/ui/
-```
-
-Your Swagger definition lives here:
-
-```
-http://localhost:8888/v1/swagger.json
-```
-
-To launch the integration tests, use tox:
-```
-sudo pip install tox
-tox
-```
-
-## Running with Docker
-
-To run the server on a Docker container, please execute the following from the root directory:
-
-```bash
-# building the image
-docker build -t swagger_server .
-
-# starting up a container
-docker run -p 8888:8888 swagger_server
-```
