--- src/api/server/requirements.txt
@@ -1,5 +1,0 @@
-connexion == 2.6.0
-swagger-ui-bundle
-python_dateutil == 2.6.0
-setuptools >= 21.0.0
-redis
