--- src/api/server/setup.py
@@ -1,33 +1,0 @@
-# coding: utf-8
-
-import sys
-from setuptools import setup, find_packages
-
-NAME = "swagger_server"
-VERSION = "1.0.0"
-# To install the library, run the following
-#
-# python setup.py install
-#
-# prerequisite: setuptools
-# http://pypi.python.org/pypi/setuptools
-
-REQUIRES = ["connexion"]
-
-setup(
-    name=NAME,
-    version=VERSION,
-    description="DNSHEALTH DNS authoritative name server verification tool",
-    author_email="",
-    url="",
-    keywords=["Swagger", "DNSHEALTH DNS authoritative name server verification tool"],
-    install_requires=REQUIRES,
-    packages=find_packages(),
-    package_data={'': ['swagger/swagger.yaml']},
-    include_package_data=True,
-    entry_points={
-        'console_scripts': ['swagger_server=swagger_server.__main__:main']},
-    long_description="""\
-    Tool used for checking if authoritative name servers comply with the necessary IANA guidlines.
-    """
-)
