--- src/api/server/swagger_server/models/__init__.py
@@ -1,10 +1,0 @@
-# coding: utf-8
-
-# flake8: noqa
-from __future__ import absolute_import
-# import models into model package
-from swagger_server.models.check import Check
-from swagger_server.models.inv_par import InvPar
-from swagger_server.models.result import Result
-from swagger_server.models.result_checks import ResultChecks
-from swagger_server.models.token import Token
