--- src/api/server/test-requirements.txt
@@ -1,6 +1,0 @@
-flask_testing==0.6.1
-coverage>=4.0.3
-nose>=1.3.7
-pluggy>=0.3.1
-py>=1.4.31
-randomize>=0.13
