--- src/api/server/tox.ini
@@ -1,10 +1,0 @@
-[tox]
-envlist = py35
-
-[testenv]
-deps=-r{toxinidir}/requirements.txt
-     -r{toxinidir}/test-requirements.txt
-
-commands=
-   nosetests \
-      []
