+++ web_api/encoder.py
@@ -1,7 +1,7 @@
 from connexion.apps.flask_app import FlaskJSONEncoder
 import six
 
-from swagger_server.models.base_model_ import Model
+from web_api.models.base_model_ import Model
 
 
 class JSONEncoder(FlaskJSONEncoder):
