+++ web_api/models/__init__.py
@@ -1,0 +1,10 @@
+# coding: utf-8
+
+# flake8: noqa
+from __future__ import absolute_import
+# import models into model package
+from web_api.models.check import Check
+from web_api.models.inv_par import InvPar
+from web_api.models.result import Result
+from web_api.models.result_checks import ResultChecks
+from web_api.models.token import Token
