+++ web_api/models/check.py
@@ -5,9 +5,9 @@ from datetime import date, datetime  # noqa: F401
 
 from typing import List, Dict  # noqa: F401
 
-from swagger_server.models.base_model_ import Model
-from swagger_server.models.token import Token  # noqa: F401,E501
-from swagger_server import util
+from web_api.models.base_model_ import Model
+from web_api.models.token import Token  # noqa: F401,E501
+from web_api import util
 
 
 class Check(Model):
