+++ web_api/models/result_checks.py
@@ -5,8 +5,8 @@ from datetime import date, datetime  # noqa: F401
 
 from typing import List, Dict  # noqa: F401
 
-from swagger_server.models.base_model_ import Model
-from swagger_server import util
+from web_api.models.base_model_ import Model
+from web_api import util
 
 
 class ResultChecks(Model):
